+++ src/models/AuthorityDocument.ts
