+++ src/models/DeathManner.ts
