+++ src/models/Deceased.ts
