+++ src/models/Inquest.ts
